fixed small error. singleChecker.js was failing to find "form1" in the page document, so I changed it to look for "document.forms[0]" instead knowing there is only one form on the page. Furthermore I wrapped a loop around the bulk of "main.sh" so that it now executes at a regular interval without needing cron or whatnot. It has a configurable sleep time, currently set to 20 seconds. Since I will be regularly polling I decided to first check if the files exist (they have all now been timestamped). If they exist, a message is printed saying so and the script moves on. No need to bother the already burdened CNATRA servers with unnecessary requests. The date variable definition is still inside the loop so at midnight it will change to the following day and then resume polling the server. Low Priority TODO: change this since we know the schedule will not be posted until much later in the day, polling at midnight is useless. Higher Priority TODO: change this to not move to the next day if the current day schedule has not yet been found (up to some limit)."

+++ main.sh
@@ -1,26 +1,70 @@
 #!/bin/bash
-echo '************ WebSkedsAutoChecker ************'
-echo '              by Alexander Buck              '
 echo ''
+echo '************ WebSkedsAutoChecker ************'
+echo '*             by Alexander Buck             *'
+echo '*********************************************'
 
 FPDIR='./FrontPage'
 
-SQUADRON='VT-6$'
+SQUADRON='VT-6'
 NAME='buck'
+SLEEPTIME=20
 declare -i JULIAN
 declare -i CALDATE
-JULIAN=$(date -v+1d +%j)
-# Conversion: 5245 == Julian Date 132 (May 12th, 2014)
-CALDATE=JULIAN+5113
-DATESTR=`date -v+1d +%Y-%m-%d`
 
-echo '** Downloading the FrontPage.%n'
-URL='http://www.cnatra.navy.mil/scheds/tw5/SQ-VT-6/$'$DATESTR'$'$SQUADRON'Frontpage.pdf'
-curl -o $FPDIR'/$'$DATESTR'$'$SQUADRON'Frontpage.pdf' $URL
 
-echo '** Downloading Schedule.%n'
-phantomjs singleCheck.js $JULIAN $NAME
+while : ; do
 
-echo '** Moving Files to Google Drive for Sharing'
-cp -R ./FrontPage '/Users/alexanderbuck/Google Drive/WebSchedule/'
-cp -R ./PNGs '/Users/alexanderbuck/Google Drive/WebSchedule/'
+  JULIAN=$(date -v+1d +%j)
+  # Conversion: 5245 == Julian Date 132 (May 12th, 2014)
+  CALDATE=JULIAN+5113
+  DATESTR=`date -v+1d +%Y-%m-%d`
+
+  if [ -f "$FPDIR/\$$DATESTR\$$SQUADRON\$Frontpage.pdf" ]
+  then
+    echo '++ Front page already downloaded.'
+  else
+    echo '**'
+    echo '** Downloading the front page.'
+    echo '**'
+    URL='http://www.cnatra.navy.mil/scheds/tw5/SQ-VT-6/$'$DATESTR'$'$SQUADRON'Frontpage.pdf'
+    curl -o $FPDIR'/$'$DATESTR'$'$SQUADRON'Frontpage.pdf' $URL
+    OUT=$?
+    if [ $OUT -eq 0 ];then
+       echo "++ Successfully downloaded front page."
+    else
+       echo "xx"
+       echo "xx Failed to download front page."
+       echo "xx"
+    fi
+  fi
+
+
+  if [  -f "./PNGs/page3_$CALDATE.png"  -a  -f "./PNGs/page4_$CALDATE.png" ]; then
+    echo "++ Schedule already downloaded."
+  else
+    echo '**'
+    echo '** Downloading schedule.'
+    echo '**'
+    phantomjs singleCheck.js $JULIAN $NAME
+
+    OUT=$?
+    if [ $OUT -eq 0 ];then
+       echo "++ Successfully downloaded schedule."
+       echo '++ Copying files to google drive for sharing.'
+       cp -R ./FrontPage '/Users/alexanderbuck/Google Drive/WebSchedule/'
+       cp -R ./PNGs '/Users/alexanderbuck/Google Drive/WebSchedule/'
+    else
+       echo "xx"
+       echo "xx Failed to download schedule."
+       echo "xx"
+    fi
+  fi
+
+  echo "It is now: `date`"
+  echo "Sleeping for $SLEEPTIME"
+  echo -n "zzz... "
+  sleep $SLEEPTIME
+  echo " ...yawn"
+
+done
